refactor: member.java 위치 변경
